+++ Assets/Scripts/Enemy/EnemyBT.cs
@@ -14,8 +14,8 @@ public class EnemyBT : MonoBehaviour
     ThirdPersonController _tpc;
 
     [Header("HP")]
-    [SerializeField] float currentHp;
-    [SerializeField] float maxhp = 15;
+    [SerializeField] int currentHp;
+    [SerializeField] int maxhp = 15;
 
     [Header("Distance")]
     [SerializeField] float moveSpeed = 6f; // Enemy 이동 속도
@@ -66,7 +66,7 @@ public class EnemyBT : MonoBehaviour
     const string _MOVE_ANIM_TRIGGER_NAME = "move";
 
     // 데미지 임시 변수
-    public float temporaryDamage = 0;
+    public int temporaryDamage = 0;
 
     bool isMove = false;
     bool isReturn = false;
